Refactored testing code
